requirements.txt
arrow==1.2.3
asgiref==3.5.2
Babel==2.12.1
binaryornot==0.4.4
certifi==2022.6.15
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.5
Django==4.1
django-browser-reload==1.6.0
django-debug-toolbar==4.1.0
django-debugtools==2.0
django-filter==23.2
django-money==3.2.0
django-oauth-toolkit==2.3.0
django-rest-paystack==2.2.2
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
filelock==3.8.0
idna==3.4
image==1.5.33
Jinja2==3.1.2
jinja2-time==0.2.0
jwcrypto==1.5.0
MarkupSafe==2.1.1
mutagen==1.46.0
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
Pillow==9.4.0
pipenv==2022.8.19
platformdirs==2.5.2
psycopg2==2.9.6
py-moneyed==3.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2022.2
urllib3==1.26.13
virtualenv==20.16.3
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wrapt==1.15.0
